chore(trumbowyg): remove trumbowyg, correctly set readOnly on braft
