+++ .prettierrc
@@ -1,5 +1,8 @@
-printWidth: 100
-singleQuote: true
-trailingComma: es5
-useTabs: true
-endOfLine: auto
+{
+	"useTabs": true,
+	"tabWidth": 4,
+	"trailingComma": "es5",
+	"singleQuote": true,
+	"endOfLine": "auto",
+	"printWidth": 100
+}
