+++ .storybook/webpack.config.js
@@ -7,22 +7,13 @@ module.exports = ({ config }) => {
 			{
 				loader: require.resolve('babel-loader'),
 				options: {
-					presets: [require.resolve('babel-preset-react-app')]
-				}
+					presets: [require.resolve('babel-preset-react-app')],
+				},
 			},
-		]
+		],
 	});
 
 	config.resolve.extensions.push('.ts', '.tsx');
 
-	config.plugins.push(
-		new webpack.ProvidePlugin({
-			$: "jquery",
-			jQuery: "jquery"
-		})
-	);
-
 	return config;
 };
-
-
