+++ package.json
@@ -39,7 +39,6 @@
     "@storybook/preset-create-react-app": "^2.1.1",
     "@types/autosize": "^3.0.6",
     "@types/classnames": "^2.2.7",
-    "@types/jquery": "^3.3.30",
     "@types/marked": "^0.6.5",
     "@types/node": "12.0.2",
     "@types/raf": "^3.4.0",
@@ -52,7 +51,6 @@
     "classnames": "^2.2.6",
     "date-fns": "^2.0.0-beta.1",
     "flowplayer-files": "^2.5.1",
-    "jquery": "^3.5.0",
     "lodash-es": "^4.17.15",
     "marked": "^0.7.0",
     "moment": "^2.24.0",
@@ -61,7 +59,6 @@
     "react-dom": "^16.8.6",
     "react-range": "^1.6.0",
     "react-select": "^3.1.0",
-    "trumbowyg": "2.21.0",
     "snyk": "^1.316.1"
   },
   "devDependencies": {
