--- src/components/WYSIWYG2/WYSIWYG2.scss
@@ -1,53 +1,0 @@
-@import '../../styles/settings/colors';
-
-.c-content table td {
-	border: 1px solid $color-gray-100;
-}
-
-.c-rich-text-editor {
-	border: 1px solid #bac7d1;
-	border-radius: 0.3rem;
-
-	&.disabled {
-		.bf-controlbar {
-			pointer-events: none;
-		}
-
-		background: $color-gray-100;
-	}
-
-	ul, ol {
-		margin: 0;
-		padding: 5px;
-	}
-
-	.bf-table-dropdown-control .button {
-		width: auto;
-	}
-
-	.bf-controlbar {
-		background: #ecf0f1;
-	}
-
-	.bf-content {
-		font-size: 15px;
-	}
-
-	.bf-controlbar .separator-line {
-		margin: 0 5px;
-		height: 35px;
-	}
-
-	.bf-controlbar .control-item {
-		margin: 0 0 0 3px;
-	}
-
-	.public-DraftEditor-content {
-		word-break: normal;
-
-		// The same as: .c-content p
-		> div > div {
-			margin: 1.5rem 0;
-		}
-	}
-}
