+++ src/components/index.ts
@@ -72,14 +72,13 @@ import { ToolbarPropsSchema } from './Toolbar/Toolbar';
 import { ToolbarItemPropsSchema } from './Toolbar/ToolbarItem/ToolbarItem';
 import { ToolbarTitlePropsSchema } from './Toolbar/ToolbarTitle/ToolbarTitle';
 import { TooltipPropsSchema } from './Tooltip/Tooltip';
-import { WYSIWYGPropsSchema } from './WYSIWYG/WYSIWYG';
 import {
 	RichEditorStateSchema,
-	WYSIWYG2MediaSchema,
-	WYSIWYG2PropsSchema,
-	WYSIWYG2ControlSchema,
-	WYSIWYG2UploadInfoSchema,
-} from './WYSIWYG2/WYSIWYG2';
+	WYSIWYGMediaSchema,
+	WYSIWYGPropsSchema,
+	WYSIWYGControlSchema,
+	WYSIWYGUploadInfoSchema,
+} from './WYSIWYG/WYSIWYG';
 
 export { Accordion } from './Accordion/Accordion';
 export type AccordionProps = AccordionPropsSchema;
@@ -310,10 +309,7 @@ export { TooltipContent, TooltipTrigger } from './Tooltip/Tooltip.slots';
 
 export { WYSIWYG } from './WYSIWYG/WYSIWYG';
 export type WYSIWYGProps = WYSIWYGPropsSchema;
-
-export { WYSIWYG2 } from './WYSIWYG2/WYSIWYG2';
-export type WYSIWYG2Props = WYSIWYG2PropsSchema;
-export type WYSIWYG2Control = WYSIWYG2ControlSchema;
+export type WYSIWYGControl = WYSIWYGControlSchema;
 export type RichEditorState = RichEditorStateSchema;
-export type WYSIWYG2Media = WYSIWYG2MediaSchema;
-export type WYSIWYG2UploadInfo = WYSIWYG2UploadInfoSchema;
+export type WYSIWYGMedia = WYSIWYGMediaSchema;
+export type WYSIWYGUploadInfo = WYSIWYGUploadInfoSchema;
